cleanup and actually fix encoding/resolution
