+++ src/__tests__/encrypt.test.ts
@@ -1,4 +1,4 @@
-import { Decrypt, Encrypt } from "../../src/utils/parse";
+import { Decrypt, Encrypt } from "../../src/utils/encrypt";
 
 const secretKey = "secretKey";
